refactor: change method paths and cleanup
