+++ backend/src/handlers/poker-tables/join/gen/errors/PlayerAlreadySeatedError.ts
