+++ backend/src/handlers/poker-tables/join/gen/errors/PlayerNotFoundAtTableError.ts
