+++ backend/src/handlers/poker-tables/join/gen/errors/SeatTakenError.ts
