+++ backend/src/handlers/poker-tables/leave/gen/errors/PlayerNotFoundAtTableError.ts
