+++ backend/src/handlers/poker-tables/leave/gen/errors/PokerTableDoesNotExistError.ts
