+++ backend/src/handlers/users/signin/gen/AbstractUsersSigninHandler.ts
@@ -4,12 +4,10 @@
 import { ResultError, ResultSuccess } from '@infra/Result';
 import { UsersSigninPayload, } from '@shared/api/gen/users/types/UsersSignin';
 
-import { AbstractUsersSigninHandler } from './gen/signinAbstractUsersSigninHandler';
+import { AbstractUsersSigninHandler } from './gen/AbstractUsersSigninHandler';
 
-import {UsernameNotFoundError} from './gen/signin/errors/UsernameNotFoundError'
-
-import {PasswordInvalidError} from './gen/signin/errors/PasswordInvalidError'
-
+import { UsernameNotFoundError } from './gen/errors/UsernameNotFoundError';
+import { PasswordInvalidError } from './gen/errors/PasswordInvalidError';
 
 export class UsersSigninHandler extends AbstractUsersSigninHandler {
   protected async getResult(payload: UsersSigninPayload) {
@@ -20,7 +18,7 @@ export class UsersSigninHandler extends AbstractUsersSigninHandler {
 import { UsersSigninOutputSchema, UsersSigninPayloadSchema } from '@shared/api/gen/users/schemas/UsersSigninSchemas';
 import { UsersSigninErrorCodes, UsersSigninOutput, UsersSigninPayload } from '@shared/api/gen/users/types/UsersSignin';
 
-import { BaseHandler } from '../../BaseHandler';
+import { BaseHandler } from '../../../BaseHandler';
 
 export abstract class AbstractUsersSigninHandler extends BaseHandler<UsersSigninPayload, UsersSigninOutput> {
   constructor() {
